+++ proto-lens-etcd/src/Proto/API/Etcd.hs
@@ -1,0 +1,5 @@
+module Proto.API.Etcd (
+    module Proto.Etcd.Api.Etcdserverpb.Rpc,
+) where
+
+import Proto.Etcd.Api.Etcdserverpb.Rpc